requirements.txt
playwright>=1.42.0
boto3>=1.34.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
pandas>=2.2.0